requirements.txt
aiomysql==0.3.2
alembic==1.18.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
bidict==0.23.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
defusedxml==0.7.1
et_xmlfile==2.0.0
granian==2.7.1
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
openpyxl==3.1.5
packaging==25.0
pillow==12.1.1
platformdirs==4.7.0
playwright==1.58.0
pluggy==1.6.0
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.11.0
PyMySQL==1.1.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-mock==3.15.1
python-dotenv==1.2.1
python-engineio==4.13.1
python-multipart==0.0.22
python-socketio==5.16.1
redis==7.1.1
reflex==0.8.26
reflex-hosting-cli==0.1.61
reportlab==4.4.10
rich==14.3.2
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.46
sqlmodel==0.0.33
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
watchfiles==1.1.1
wrapt==2.1.1
wsproto==1.3.2
